third_party/WebKit/LayoutTests/http/tests/devtools/network/font-face-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that a used font-face is reported and an unused font-face is not reported.

RequestStarted: font-face.html
RequestFinished: font-face.html
RequestStarted: used.ttf
RequestFinished: used.ttf
